ICU-21757 Stop sharing `utilities-for-cldr`
Moves UnicodeMap and related classes to core.
Also removes `CollectionUtilities`, `UOption`, and `ElapsedTimer`.
They will end up in UnicodeTools, CLDR, and CLDR respectively.
