+++ icu4j/main/core/src/test/java/com/ibm/icu/dev/test/UnicodeMapBoilerplateTest.java
@@ -6,23 +6,24 @@
  * others. All Rights Reserved.                                                *
  *******************************************************************************
  */
-package com.ibm.icu.dev.test.translit;
+package com.ibm.icu.dev.test;
 
 import java.util.List;
 
 import org.junit.Test;
 
-import com.ibm.icu.dev.test.TestBoilerplate;
-import com.ibm.icu.text.UnicodeSet;
+import com.ibm.icu.impl.UnicodeMap;
 
 /**
  * Moved from UnicodeMapTest
  */
-public class UnicodeSetBoilerplateTest extends TestBoilerplate<UnicodeSet> {
+public class UnicodeMapBoilerplateTest extends TestBoilerplate<UnicodeMap> {
 
-    public void TestUnicodeSetBoilerplate() throws Exception {
+    private static String[] TEST_VALUES = {"A", "B", "C", "D", "E", "F"};
+
+    public void TestUnicodeMapBoilerplate() throws Exception {
     }
- 
+
     @Test
     public void test() throws Exception {
         _test();
@@ -31,8 +32,7 @@ public class UnicodeSetBoilerplateTest extends TestBoilerplate<UnicodeSet> {
     /* (non-Javadoc)
      * @see com.ibm.icu.dev.test.TestBoilerplate#_hasSameBehavior(java.lang.Object, java.lang.Object)
      */
-    @Override
-    protected boolean _hasSameBehavior(UnicodeSet a, UnicodeSet b) {
+    protected boolean _hasSameBehavior(UnicodeMap a, UnicodeMap b) {
         // we are pretty confident in the equals method, so won't bother with this right now.
         return true;
     }
@@ -40,14 +40,16 @@ public class UnicodeSetBoilerplateTest extends TestBoilerplate<UnicodeSet> {
     /* (non-Javadoc)
      * @see com.ibm.icu.dev.test.TestBoilerplate#_addTestObject(java.util.List)
      */
-    @Override
-    protected boolean _addTestObject(List<UnicodeSet> list) {
-        if (list.size() > 32) return false;
-        UnicodeSet result = new UnicodeSet();
+    protected boolean _addTestObject(List<UnicodeMap> list) {
+        if (list.size() > 30) return false;
+        UnicodeMap result = new UnicodeMap();
         for (int i = 0; i < 50; ++i) {
-            result.add(random.nextInt(100));
+            int start = random.nextInt(25);
+            String value = TEST_VALUES[random.nextInt(TEST_VALUES.length)];
+            result.put(start, value);
         }
         list.add(result);
         return true;
     }
+
 }
