+++ icu4j/main/translit/src/test/java/com/ibm/icu/dev/test/translit/TransliteratorDisorderedMarksTest.java
@@ -3,6 +3,8 @@
 
 package com.ibm.icu.dev.test.translit;
 
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Map.Entry;
 
 import org.junit.AfterClass;
@@ -11,7 +13,6 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import com.ibm.icu.dev.test.TestFmwk;
-import com.ibm.icu.dev.util.UnicodeMap;
 import com.ibm.icu.text.CanonicalIterator;
 import com.ibm.icu.text.Normalizer2;
 import com.ibm.icu.text.Transliterator;
@@ -58,8 +59,8 @@ public class TransliteratorDisorderedMarksTest extends TestFmwk {
         //        logln("NFKD Source: " + nfkdSource.toPattern(false));
         //        logln("NFKD Target: " + nfkdTarget.toPattern(false));
 
-        UnicodeMap<UnicodeSet> leadToTrail = new UnicodeMap();
-        UnicodeMap<UnicodeSet> leadToSources = new UnicodeMap();
+        Map<Integer, UnicodeSet> leadToTrail = new HashMap<>();
+        Map<Integer, UnicodeSet> leadToSources = new HashMap<>();
         UnicodeSet nonStarters = new UnicodeSet("[:^ccc=0:]").freeze();
         CanonicalIterator can = new CanonicalIterator("");
 
@@ -101,8 +102,8 @@ public class TransliteratorDisorderedMarksTest extends TestFmwk {
         }
 
 
-        for (Entry<String, UnicodeSet> x : leadToSources.entrySet()) {
-            String lead = x.getKey();
+        for (Entry<Integer, UnicodeSet> x : leadToSources.entrySet()) {
+            Integer lead = x.getKey();
             UnicodeSet sources = x.getValue();
             UnicodeSet trailSet = leadToTrail.get(lead);
             for (String source : sources) {
