+++ icu4j/pom.xml
@@ -44,7 +44,6 @@
     <module>demos</module>
     <module>samples</module>
     <module>tools/misc</module>
-    <module>tools/utilities-for-cldr</module>
     <module>perf-tests</module>
   </modules>
